\name{scTK_launch}
\title{Launch singleCellTK in its Shiny interface}
\usage{
  scTK_launch <- function(external_browser = TRUE)
}
\arguments{
  \item{external_browser}{Use the default external web browser rather than the one included in Rstudio}
\description{
  This function launches the Shiny interface of singleCellTK, possibly in the default system web browser.
}
\author{
  Bastien Job
}
